working resources- writing for ProfileDetailsTable.txt
