+++ src/main/java/com/facebook/ui/MyProfileUI.java
@@ -1,6 +1,7 @@
 package com.facebook.ui;
 
 import com.facebook.model.User;
+import com.facebook.service.ShowProfileDetailsService;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -50,7 +51,7 @@ public class MyProfileUI extends UI {
                     break;
                 case "1":
                     System.out.println("Show Profile Details");
-                    MyProfileUI.showProfileDetails();
+                    ShowProfileDetailsService.showProfileDetails();
                     break;
                 case "2":
                     System.out.println("Edit Profile Details");
@@ -70,47 +71,47 @@ public class MyProfileUI extends UI {
         }
     }
 
-    public static void showProfileDetails() throws IOException {
-        Scanner in = new Scanner(System.in);
+//    public static void showProfileDetails() throws IOException {
+//        Scanner in = new Scanner(System.in);
+//
+//
+//        String name = "not assigned";
+//        int age = 0;
+////        String email="not assigned";
+//        String sex = "not assigned";
+//        /*String livesIn = "not assigned";
+////        String fromTown = "not assigned";
+////        String worksAt = "not assigned";
+////        String studies = "not assigned";
+////        String relationshipStatus = "not assigned";
+//
+//         */
+//        int friendsNo = 0;
+//        //add date of join
+//
+//// Read profile from file
+//        int id = User.getId();
+////        System.out.println("ID="+ id);
+//        boolean idCheck = false;
+//        Path profileDetailsTable = Paths.get("src\\main\\resources", "CurrentUserIDTable.txt");
+//        List<String> lines = Files.readAllLines(profileDetailsTable);
+//        String [] userDetails;
+//        for (String line : lines) {
+//            userDetails = line.split(";");
+//            if (userDetails[0].equals(Integer.toString(id))) {
+//                System.out.println("----------------------------MY PROFILE ---------------------------");
+//                System.out.println("Name = "+ userDetails[3] + "\nAge = " + userDetails[4] + "\nSex = " + userDetails[5]);
+//
+//            }
+//        }
+//
 
 
-        String name = "not assigned";
-        int age = 0;
-//        String email="not assigned";
-        String sex = "not assigned";
-        /*String livesIn = "not assigned";
-//        String fromTown = "not assigned";
-//        String worksAt = "not assigned";
-//        String studies = "not assigned";
-//        String relationshipStatus = "not assigned";
 
-         */
-        int friendsNo = 0;
-        //add date of join
-
-// Read profile from file
-        int id = User.getId();
-//        System.out.println("ID="+ id);
-        boolean idCheck = false;
-        Path profileDetailsTable = Paths.get("src\\main\\resources", "CurrentUserIDTable.txt");
-        List<String> lines = Files.readAllLines(profileDetailsTable);
-        String [] userDetails;
-        for (String line : lines) {
-            userDetails = line.split(";");
-            if (userDetails[0].equals(Integer.toString(id))) {
-                System.out.println("----------------------------MY PROFILE ---------------------------");
-                System.out.println("Name = "+ userDetails[3] + "\nAge = " + userDetails[4] + "\nSex = " + userDetails[5]);
-
-            }
-        }
-
-
-
-
-        System.out.println("\n Name= " + name); // still to create regex
-        System.out.println("\n Age= " + age + " years");  //still create regex - age at account creation must be 18 or higher
+//        System.out.println("\n Name= " + name); // still to create regex
+//        System.out.println("\n Age= " + age + " years");  //still create regex - age at account creation must be 18 or higher
 //        System.out.println("\n Name= " +email);
-        System.out.println("\n Sex= " + sex);  // still to Create enum
+//        System.out.println("\n Sex= " + sex);  // still to Create enum
 //        System.out.println("\n Name= " + livesIn);
 //        System.out.println("\n Name= " + fromTown);
 //        System.out.println("\n Name= " + worksAt);
@@ -119,7 +120,7 @@ public class MyProfileUI extends UI {
 //        System.out.println("\n Follow By= " + friendsNo + " people");// still to create counter
 
 
-    }
+//    }
 
     @Override
     public void closeApplication() {
