+++ src/main/resources/CurrentUserIDTable.txt
@@ -1,14 +1,1 @@
-1; radu@mail.com; radupassword; Radu Baciu; 22; male
-2; mula@mail.com; mulapassword; Mula Mare; 44; male
-3; gigi@mail.com; gigipassword; Gigi Becali; 33; male
-4; a@mail.com; apass; A Name; 22; female
-5; b@mail.com; bpass; B Name; 33; male
-6; c@mail.com; cpass; C Name; 44; male
-7; d@mail.com; dpass; D Name; 44; male
-8; e@mail.com; 2pass; E Name; 44; male
-9; f; fpass; F Name; 44; male
-10; g@mail.com; gpass; G Name; 22; male
-11; h@mail.com; hpass; H Name; 44; male
-12; i@mail.com; ipass; I Name; 33; male
-13; j@mail.com; jpass; J Name; 33; male
-14; k@mail.com; kpass; K Name; 44; male
+18;19;
